+++ src/main/java/com/barlinc/unusual_prehistory/client/models/entity/MetriorhynchusModel.java
@@ -112,11 +112,12 @@ public class MetriorhynchusModel extends UP2Model<Metriorhynchus> {
 	@Override
 	public void setupAnim(Metriorhynchus entity, float limbSwing, float limbSwingAmount, float ageInTicks, float netHeadYaw, float headPitch) {
 		this.root().getAllParts().forEach(ModelPart::resetPose);
+        float deg = ((float) Math.PI / 180F);
 
 		if (!entity.isInWater()) {
 			this.animateWalk(MetriorhynchusAnimations.WALK, limbSwing, limbSwingAmount, 2, 4);
 		} else if (entity.getPose() != UP2Poses.DEATH_ROLL.get()) {
-            this.swim_control.xRot = headPitch * ((float) Math.PI / 180F);
+            this.swim_control.xRot = headPitch * deg;
             this.animateWalk(entity.isRunning() ? MetriorhynchusAnimations.SWIMFAST : MetriorhynchusAnimations.SWIM, limbSwing, limbSwingAmount, 1, 2);
         }
 
@@ -130,8 +131,8 @@ public class MetriorhynchusModel extends UP2Model<Metriorhynchus> {
         this.animate(entity.deathRoll2AnimationState, MetriorhynchusAnimations.DEATHROLL2, ageInTicks);
         this.animate(entity.bellowAnimationState, MetriorhynchusAnimations.BELLOW_BLEND, ageInTicks);
 
-		this.head.xRot += (headPitch * ((float) Math.PI / 180)) / 2;
-		this.head.yRot += (netHeadYaw * ((float) Math.PI / 180)) / 2;
+		this.head.xRot += (headPitch * deg) / 2;
+		this.head.yRot += (netHeadYaw * deg) / 2;
 	}
 
 	@Override