+++ src/main/java/com/barlinc/unusual_prehistory/datagen/UP2LanguageProvider.java
@@ -86,6 +86,7 @@ public class UP2LanguageProvider extends LanguageProvider {
         this.sound(UP2SoundEvents.CARNOTAURUS_BITE, "Carnotaurus bites");
         this.sound(UP2SoundEvents.CARNOTAURUS_CHARGE, "Carnotaurus charges");
         this.sound(UP2SoundEvents.CARNOTAURUS_HEADBUTT, "Carnotaurus headbutts");
+        this.sound(UP2SoundEvents.CARNOTAURUS_ROAR, "Carnotaurus roars");
 
         this.sound(UP2SoundEvents.DIPLOCAULUS_HURT, "Diplocaulus hurts");
         this.sound(UP2SoundEvents.DIPLOCAULUS_DEATH, "Diplocaulus dies");