+++ src/main/java/com/barlinc/unusual_prehistory/entity/ai/goals/megalania/MegalaniaAttackGoal.java
@@ -1,5 +1,6 @@
 package com.barlinc.unusual_prehistory.entity.ai.goals.megalania;
 
+import com.barlinc.unusual_prehistory.entity.Carnotaurus;
 import com.barlinc.unusual_prehistory.entity.Megalania;
 import com.barlinc.unusual_prehistory.entity.ai.goals.AttackGoal;
 import com.barlinc.unusual_prehistory.entity.utils.UP2Poses;
@@ -7,15 +8,17 @@ import com.barlinc.unusual_prehistory.registry.UP2SoundEvents;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.Pose;
+import net.minecraft.world.entity.ai.attributes.Attributes;
+import net.minecraft.world.entity.ai.targeting.TargetingConditions;
+import net.minecraft.world.entity.monster.Creeper;
 import net.minecraft.world.entity.player.Player;
 
+import java.util.List;
 import java.util.Objects;
 
 public class MegalaniaAttackGoal extends AttackGoal {
 
     private final Megalania megalania;
-    private int tailWhipCooldown;
-    private int biteCooldown;
 
     public MegalaniaAttackGoal(Megalania megalania) {
         super(megalania);
@@ -29,20 +32,6 @@ public class MegalaniaAttackGoal extends AttackGoal {
     }
 
     @Override
-    public void start() {
-        super.start();
-        this.tailWhipCooldown =0;
-        this.biteCooldown = 0;
-    }
-
-    @Override
-    public void stop() {
-        super.stop();
-        this.tailWhipCooldown =0;
-        this.biteCooldown = 0;
-    }
-
-    @Override
     public void tick() {
         LivingEntity target = this.megalania.getTarget();
         if (target != null) {
@@ -52,31 +41,45 @@ public class MegalaniaAttackGoal extends AttackGoal {
             double distance = this.megalania.distanceToSqr(target.getX(), target.getY(), target.getZ());
             int attackState = this.megalania.getAttackState();
 
-            if (this.megalania.isInWater()) {
-                this.megalania.getNavigation().moveTo(target, 1.2D);
-            } else if (this.megalania.getTemperatureState().equals(Megalania.TemperatureStates.NETHER)) {
-                this.megalania.getNavigation().moveTo(target, 1.8D);
-            } else {
-                this.megalania.getNavigation().moveTo(target, 2.0D);
-            }
-
-            if (this.biteCooldown > 0) biteCooldown--;
-            if (this.tailWhipCooldown > 0 && !megalania.isInWater()) tailWhipCooldown--;
-
             if (attackState == 1) {
+                this.megalania.getNavigation().moveTo(target, this.megalania.isInWater() ? this.getSpeedMultiplier() * 0.5F : this.getSpeedMultiplier() * 0.7F);
                 this.tickBite();
-            } else if (attackState == 2) {
+            }
+            else if (attackState == 2) {
                 this.tickTailWhip();
-            } else {
+            }
+            else {
+                this.megalania.getNavigation().moveTo(target, this.megalania.isInWater() ? this.getSpeedMultiplier() * 0.7F : this.getSpeedMultiplier());
                 if (distance <= this.getAttackReachSqr(target)) {
-                    if (!this.megalania.isInWater()) {
-                        if (this.megalania.getRandom().nextFloat() < 0.38F) this.megalania.setAttackState(2);
-                        else this.megalania.setAttackState(1);
-                    } else {
-                        this.megalania.setAttackState(1);
-                    }
+                    this.selectAttack();
                 }
             }
+        }
+    }
+
+    private double getSpeedMultiplier() {
+        switch (megalania.getTemperatureState()) {
+            case COLD -> {
+                return 1.4D;
+            }
+            case WARM -> {
+                return 1.9D;
+            }
+            case NETHER -> {
+                return 1.5D;
+            }
+            default -> {
+                return 2.0D;
+            }
+        }
+    }
+
+    private void selectAttack() {
+        if (megalania.isInWater()) {
+            if (megalania.biteCooldown == 0) this.megalania.setAttackState(1);
+        } else {
+            if (megalania.getRandom().nextBoolean() && megalania.talWhipCooldown == 0) this.megalania.setAttackState(2);
+            else if (megalania.biteCooldown == 0) this.megalania.setAttackState(1);
         }
     }
 
@@ -92,9 +95,8 @@ public class MegalaniaAttackGoal extends AttackGoal {
                 this.megalania.applyPoison(target);
             }
         }
-        if (this.timer > 22) {
+        if (this.timer > 20) {
             this.timer = 0;
-            this.biteCooldown = 10 + megalania.getRandom().nextInt(6);
             this.megalania.setAttackState(0);
         }
     }
@@ -102,25 +104,30 @@ public class MegalaniaAttackGoal extends AttackGoal {
     private void tickTailWhip() {
         this.timer++;
         this.megalania.getNavigation().stop();
-        LivingEntity target = this.megalania.getTarget();
 
         if (timer == 1) this.megalania.setPose(UP2Poses.TAIL_WHIPPING.get());
         if (timer == 9) this.megalania.playSound(UP2SoundEvents.MEGALANIA_TAIL_SWING.get(), 1.0F, 1.0F);
         if (this.timer == 12) this.megalania.addDeltaMovement(this.megalania.getLookAngle().scale(2.0D).multiply(0.25D, 0, 0.25D));
         if (this.timer == 14) {
-            if (this.megalania.distanceTo(Objects.requireNonNull(target)) <= this.getAttackReachSqr(target)) {
-                this.megalania.swing(InteractionHand.MAIN_HAND);
-                this.megalania.doHurtTarget(target);
-                this.megalania.strongKnockback(target, 1.3D, 0.1D);
-                if (target.isDamageSourceBlocked(this.megalania.damageSources().mobAttack(this.megalania)) && target instanceof Player player){
+            this.hurtNearbyEntities();
+        }
+        if (this.timer > 30) {
+            this.timer = 0;
+            this.megalania.setAttackState(0);
+        }
+    }
+
+    private void hurtNearbyEntities() {
+        List<LivingEntity> nearbyEntities = megalania.level().getNearbyEntities(LivingEntity.class, TargetingConditions.forCombat(), megalania, megalania.getBoundingBox().inflate(2.5, -0.5, 2.5));
+        if (!nearbyEntities.isEmpty()) {
+            nearbyEntities.stream().filter(entity -> !entity.is(megalania) && !entity.isAlliedTo(megalania)).limit(3).forEach(entity -> {
+                entity.hurt(entity.damageSources().mobAttack(megalania), (float) megalania.getAttributeValue(Attributes.ATTACK_DAMAGE));
+                megalania.strongKnockback(entity, 1.3D, 0.2D);
+                if (entity.isDamageSourceBlocked(megalania.damageSources().mobAttack(megalania)) && entity instanceof Player player) {
                     player.disableShield(true);
                 }
-            }
-        }
-        if (this.timer > 28) {
-            this.timer = 0;
-            this.tailWhipCooldown = 20 + megalania.getRandom().nextInt(24);
-            this.megalania.setAttackState(0);
+                megalania.swing(InteractionHand.MAIN_HAND);
+            });
         }
     }